[Bug] Fixing bugs in responsiveness
Yes we could argue that these weren't bugs but lets be honest it was so
bad they actually were bugs. Added some tests and fixed some issues
where we would have to wait a full update cycle for our alerts to start.
Also added a synchronization point for removing alerts as part of the
alert update thread rather than deal with a race condition in all future
alerts where they could be disposed of mid-update.